I wrote a method for calculating monthly payments. I'm starting to create a method for writing the result in csv format

+++ src/test/java/CreditCalculatorTest.java
@@ -19,7 +19,7 @@ public class CreditCalculatorTest {
         creditCalculator = new CreditCalculator();
     }
 
-    public void getLoanInformation(int loanAmount, double interestRate, int duration) {
+    public void getLoanInformation(double loanAmount, double interestRate, int duration) {
         File csvFile = new File("src/test/java/testResources/present.csv");
         try {
             CSVReader csvReader = new CSVReader(new FileReader(csvFile));
